shared/pyproject.toml
[project]
name = "jib-shared"
version = "0.1.0"
description = "Shared modules for jib (enrichment, logging, notifications, etc.)"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["beads*", "enrichment*", "git_utils*", "jib_config*", "jib_logging*", "notifications*", "text_utils*"]
